neo4j/requirements.txt
neo4j~=5.28.0
python-dotenv
boto3~=1.34.0
httpx~=0.27.0
tenacity~=8.2.3
pymupdf~=1.24.9